Add carousel button actions and network error views

+++ app/src/main/java/com/simonassi/globoplay/data/Gender.kt
@@ -3,6 +3,6 @@ package com.simonassi.globoplay.data
 import com.google.gson.annotations.SerializedName
 
 data class Gender(
-    @SerializedName("id") val id: Int,
+    @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String,
 )
